Add fixes according review comments
